feat: refact listrequest and listmeta
