+++ src/core/cast_member/tests/application/integration/test_list_cast_member.py
@@ -1,7 +1,8 @@
 
 import pytest
 
-from src.core.cast_member.application.use_cases.list_cast_member import ListCastMember, ListCastMemberRequest
+from src.core._shared.list import ListRequest
+from src.core.cast_member.application.use_cases.list_cast_member import ListCastMember
 from src.core.cast_member.domain.cast_member import CastMember, CastMemberType
 from src.core.cast_member.infra.in_memory_cast_member_repository import InMemoryCastMemberRepository
 
@@ -29,7 +30,7 @@ class TestListCastMember:
 
         use_case = ListCastMember(cast_member_reṕository)
 
-        response = use_case.execute(ListCastMemberRequest())
+        response = use_case.execute(ListRequest())
 
         assert len(response.data) == 2
         assert response.data[0].name == "Bruce Willis"
@@ -42,6 +43,6 @@ class TestListCastMember:
     def test_list_no_cast_member_return_empty_list(self):
         cast_member_reṕository = InMemoryCastMemberRepository()
         use_case = ListCastMember(cast_member_reṕository)
-        response = use_case.execute(ListCastMemberRequest())
+        response = use_case.execute(ListRequest())
 
         assert response.data == []
