+++ src/django_project/_shared/serializer.py
@@ -1,0 +1,6 @@
+from rest_framework import serializers
+
+class ListOutputMetaSerializer(serializers.Serializer):
+    current_page = serializers.IntegerField()
+    per_page = serializers.IntegerField()
+    total = serializers.IntegerField()
