+++ src/django_project/genre/serializers.py
@@ -1,5 +1,8 @@
 from rest_framework import serializers
 
+from src.django_project._shared.serializer import ListOutputMetaSerializer
+
+
 class GenreOutputSerializer(serializers.Serializer):
     id = serializers.UUIDField()
     name = serializers.CharField(max_length=255)
@@ -8,6 +11,7 @@ class GenreOutputSerializer(serializers.Serializer):
 
 class ListGenreResponseSerializer(serializers.Serializer):
     data = GenreOutputSerializer(many=True)
+    meta = ListOutputMetaSerializer()
 
 class SetField(serializers.ListField):
     def to_internal_value(self, data):
